+++ app/Support/System/Updates/UpdateMigrationRunner.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace App\Support\System\Updates;
+
+use WebBlocks\Cms\Support\System\Updates\UpdateMigrationRunner as PackageUpdateMigrationRunner;
+
+class UpdateMigrationRunner extends PackageUpdateMigrationRunner {}